minor changes and cleanup
