+++ tabs/src/components/Tab.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 // import { Scheduler } from "./BeaverHours/Scheduler";
-import { Scheduler } from "./BeaverHours/SchedulerRewrite";
+import { Scheduler } from "./BeaverHours/Scheduler";
 
 export default function Tab() {
   return (
